requirements.txt
blessed==1.17.6
gitdb==4.0.5
GitPython==3.1.11
python-editor==1.0.4
readchar==2.0.1
simple-term-menu==0.10.4
six==1.15.0
smmap==3.0.4
wcwidth==0.2.5
